refactor: Rename `migrate.rs` -> `migartion.rs`
